+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## Unreleased
+- Added decimalPlaces parameter to round and floor.
+
 ## 0.0.1-dev
 - Initial